requirements.txt
pyairtable==2.3.3
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
requests==2.31.0
